requirements.txt
alembic==1.13.1
annotated-types==0.6.0
asgiref==3.7.2
blinker==1.7.0
Django==4.0.2
django-rest-enumfield==0.2.0
django-rest-witchcraft==0.12.1
django-sorcery==0.13.0
djangorestframework==3.14.0
greenlet==3.0.3
inflect==7.0.0
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==1.4.52
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
